requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
spotipy==2.23.0
python-dotenv==1.0.0
requests==2.31.0 